requirements.txt
mysqlclient
pytest
coverage
codecov
wheel
pyyaml
docopt
requests
